@@ -1,4 +1,3 @@
-export * from './point';
 export * from './interfaces';
-export * from './collection';
+export * from './query';
 export * from './client';
