@@ -1,8 +1,35 @@
+import { Point, Feature, Coordinates } from './interfaces';
+
+import turfDistance from '@turf/distance';
+import turfBearing from '@turf/bearing';
+
+
+export function distance(from: Coordinates, to: Coordinates) {
+  return turfDistance(toGeoJSONFeature(from), toGeoJSONFeature(to));
+}
+
+export function bearing(from: Coordinates, to: Coordinates) {
+  return turfBearing(toGeoJSONFeature(from), toGeoJSONFeature(to));
+}
+
+export function toGeoJSONFeature(coordinates: Coordinates, props?: any): Feature<Point> {
+  coordinates = flip(coordinates) as Coordinates;
+  return {
+    type: 'Feature',
+    geometry: {
+      type: 'Point',
+      coordinates
+    },
+    properties: props
+  };
+}
+
+
 export function flip(arr) {
   return [arr[1], arr[0]];
 }
 
-export function setPrecsion(km: number): number {
+export function setPrecision(km: number): number {
   switch (true) {
     case km <= 0.00477:
       return 9;
